Started implementing SD datalogging
At this stage, no information is stored or retrieved from the SD card,
but all the groundwork is being put in place to be able to implement it
across the system as soon as it happens.
Introduced many compiler #if SD_EN in order to keep functionality if SD
card is not desired.
Introduced the ability to call setDateTime by passing it an array
containing the date and time in the sequence sent to the RTC.